fix: cleanup ability manager
