+++ global/packages/core/src/auth/ability-manager/index.ts
@@ -5,6 +5,6 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+export * from './build';
 export * from './module';
 export * from './type';
-export * from './utils';
