--- global/packages/core/src/auth/ability-manager/utils/extend.ts
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) 2022.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import type { Ability } from '../type';
-
-export function extendAbility(ability: Ability) {
-    if (typeof ability.target === 'undefined') {
-        ability.target = null;
-    }
-
-    if (typeof ability.inverse === 'undefined') {
-        ability.inverse = null;
-    }
-
-    return ability;
-}
