+++ server/services/core/src/domains/robot/repository.ts
@@ -9,7 +9,7 @@ import type {
     Ability, Robot,
     Role,
 } from '@authup/core';
-import { buildAbilityDescriptor, createNanoID } from '@authup/core';
+import { buildAbility, createNanoID } from '@authup/core';
 
 import { compare, hash } from '@authup/server-kit';
 import type { DataSource, EntityManager } from 'typeorm';
@@ -79,7 +79,7 @@ export class RobotRepository extends Repository<RobotEntity> {
 
         const result : Ability[] = [];
         for (let i = 0; i < entities.length; i++) {
-            result.push(buildAbilityDescriptor(entities[i]));
+            result.push(buildAbility(entities[i]));
         }
 
         return result;
